Move CourtCaseDetailsSummaryBox to a new component/ file
