+++ src/features/CourtCaseDetails/CourtCaseDetails.tsx
@@ -7,67 +7,8 @@ import LinkButton from "components/LinkButton"
 import { Heading, Paragraph, Table, Tag } from "govuk-react"
 import CourtCase from "services/entities/CourtCase"
 import { displayedDateFormat } from "utils/formattedDate"
-import { createUseStyles } from "react-jss"
-import { gdsLightGrey } from "utils/colours"
 import UrgentBadge from "features/CourtCaseList/tags/UrgentBadge"
-
-interface CourtCaseDetailsSummaryBoxFieldProps {
-  label: string
-  value: string | null | undefined
-}
-
-const CourtCaseDetailsSummaryBoxField = ({ label, value }: CourtCaseDetailsSummaryBoxFieldProps) => (
-  <div>
-    <b className="govuk-!-static-padding-right-3">{label}</b> {value}
-  </div>
-)
-
-const useStyles = createUseStyles({
-  "court-case-details-summary-box": {
-    display: "grid",
-    gridTemplateColumns: "auto auto auto",
-    backgroundColor: gdsLightGrey,
-    padding: "20px 0",
-    margin: "34px 0",
-    "& div": {
-      padding: "10px",
-      fontSize: "19px",
-      lineHeight: "25px"
-    }
-  }
-})
-
-interface CourtCaseDetailsSummaryBoxProps {
-  asn: string | null
-  courtCode: string | null
-  courtName: string
-  courtReference: string
-  pnci: string | undefined
-  ptiurn: string
-}
-
-const CourtCaseDetailsSummaryBox = ({
-  asn,
-  courtCode,
-  courtName,
-  courtReference,
-  pnci,
-  ptiurn
-}: CourtCaseDetailsSummaryBoxProps) => {
-  const classes = useStyles()
-
-  return (
-    <div className={classes["court-case-details-summary-box"]}>
-      <CourtCaseDetailsSummaryBoxField label="PTIURN" value={ptiurn} />
-      <CourtCaseDetailsSummaryBoxField label="ASN" value={asn} />
-      <CourtCaseDetailsSummaryBoxField label="PNCID" value={pnci} />
-      <CourtCaseDetailsSummaryBoxField label="Court name" value={courtName} />
-      <CourtCaseDetailsSummaryBoxField label="Court code (LJA)" value={courtCode} />
-      <CourtCaseDetailsSummaryBoxField label="Court case reference" value={courtReference} />
-    </div>
-  )
-}
-
+import CourtCaseDetailsSummaryBox from "./CourtCaseDetailsSummaryBox"
 interface Props {
   courtCase: CourtCase
   aho: AnnotatedHearingOutcome
